requirements.txt
asgiref==3.8.1
Django==4.2.16
django-cors-headers==4.9.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.1
psycopg2-binary==2.9.11
PyJWT==2.10.1
setuptools==80.9.0
sqlparse==0.5.3
wheel==0.45.1
python-dotenv
gunicorn
Pillow
